Add something

+++ src/Main.java
@@ -1,8 +1,8 @@
 public class Main {
 
     public static void main(String[] args) {
-        float perday = 0.5f;
-        float pathlearning = 20f;
+        double perday = 0.5f;
+        double pathlearning = 20f;
 
         System.out.println("Days needed:" + pathlearning/perday);
     }
